--- a/Library/Homebrew/Gemfile
+++ b/Library/Homebrew/Gemfile
@@ -30,6 +30,7 @@ end
 # vendored gems
 gem "activesupport"
 gem "concurrent-ruby"
+gem "did_you_mean" # remove when HOMEBREW_REQUIRED_RUBY_VERSION >= 2.7
 gem "mechanize"
 gem "patchelf"
 gem "plist"