--- a/Library/Homebrew/brew.sh
+++ b/Library/Homebrew/brew.sh
@@ -583,13 +583,10 @@ then
 
   # Don't allow non-developers to customise Ruby warnings.
   unset HOMEBREW_RUBY_WARNINGS
-
-  # Disable Ruby options we don't need.
-  RUBY_DISABLE_OPTIONS="--disable=did_you_mean,rubyopt"
-else
-  # Don't disable did_you_mean for developers as it's useful.
-  RUBY_DISABLE_OPTIONS="--disable=rubyopt"
 fi
+
+# Disable Ruby options we don't need.
+RUBY_DISABLE_OPTIONS="--disable=rubyopt"
 
 if [[ -z "${HOMEBREW_RUBY_WARNINGS}" ]]
 then