--- a/Library/Homebrew/search.rb
+++ b/Library/Homebrew/search.rb
@@ -87,6 +87,8 @@ module Homebrew
                 .search(string_or_regex)
                 .sort
 
+      results += Formula.fuzzy_search(string_or_regex)
+
       results.map do |name|
         formula, canonical_full_name = begin
           f = Formulary.factory(name)