@MikeMcQuaid 2021-06-21
> ==> Searching for similarly named formulae...
> These similarly named formulae were found:
> ripgrep                                                                                       pipgrip
> To install one of them, run (for example):
>   brew install ripgrep
> Error: No available formula or cask with the name "ripgrip". Did you mean ripgrep or pipgrip?

Nice! Feels like the output is slightly duplicated, though. Maybe worth removing from the exception message?

> ```
>     1.00 ± 0.02 times faster than 'HOMEBREW_RUBY_DISABLE_OPTIONS=--disable=rubyopt HOMEBREW_NO_ANALYTICS=1 brew info docker'
> ```

This seems fine/worth it 👍🏻 

> * Despite it works in brew (for me), tests can't find did_you_mean gem (need to check if we use different sets of gems for brew and tests). Any suggestions are welcome here 🙂

May need to adjust the load path and/or requires in `spec_helper.rb` or `tests.rb`.