@bayandin 2021-06-22
> ==> Searching for similarly named formulae...
> These similarly named formulae were found:
> ripgrep pipgrip
> To install one of them, run (for example):
> brew install ripgrep
> Error: No available formula or cask with the name "ripgrip". Did you mean ripgrep or pipgrip?
> 
> Nice! Feels like the output is slightly duplicated, though. Maybe worth removing from the exception message?

```
==> Searching for similarly named formulae...
Error: No similarly named formulae found.
Error: No available formula or cask with the name "ripgrip".
```

Yep, it looks like now, we have similar duplication. I've removed the second message.

> > * Despite it works in brew (for me), tests can't find did_you_mean gem (need to check if we use different sets of gems for brew and tests). Any suggestions are welcome here 🙂
> 
> May need to adjust the load path and/or requires in `spec_helper.rb` or `tests.rb`.

We run tests via `bundle exec ...` and we have `BUNDLE_DISABLE_SHARED_GEMS: "true"`. It looks like this doesn't allow bundler to pick up existing (not bundled) gems, even if they're bundled with ruby.
I've decided to bundle also `did_you_mean` gem. It's pure ruby and relatively small:
```
$ scc --no-cocomo ./Library/Homebrew/vendor/bundle/ruby/2.6.0/gems/did_you_mean-1.5.0/lib/did_you_mean
───────────────────────────────────────────────────────────────────────────────
Language                 Files     Lines   Blanks  Comments     Code Complexity
───────────────────────────────────────────────────────────────────────────────
Ruby                        17       696      109       102      485         45
───────────────────────────────────────────────────────────────────────────────
Total                       17       696      109       102      485         45
───────────────────────────────────────────────────────────────────────────────
Processed 17595 bytes, 0.018 megabytes (SI)
───────────────────────────────────────────────────────────────────────────────
```
We'll be able to get rid of it when switching to ruby 2.7 in the future.

Alternatively, if we don't want to bundle it, we can use this gem opportunistically (like `return [] unless defined?(DidYouMean::SpellChecker)`)